correct code for the logical VC
