+++ quavi/quavi/Controllers/POIInfoViewController.swift
@@ -10,18 +10,6 @@ import UIKit
 
 class POIInfoViewController: UIViewController {
     
-    lazy var reachedLabel: UILabel = {
-        var label = UILabel()
-        label.text = "You've reached:"
-        return label
-    }()
-    
-    lazy var pointNameLabel: UILabel = {
-        var label = UILabel()
-        label.text = "The Empire State Building"
-        return label
-    }()
-    
     lazy var continueButton: UIButton = {
         let button = UIButton(frame: CGRect(x: 0, y: 0, width: 100, height: 100))
         button.setTitle("Next", for: .normal)
