+++ quavi/quavi/Controllers/POIPopUpViewController.swift
@@ -9,6 +9,18 @@
 import UIKit
 
 class POIPopUpViewController: UIViewController {
+    
+    lazy var reachedLabel: UILabel = {
+        var label = UILabel()
+        label.text = "You've reached:"
+        return label
+    }()
+    
+    lazy var pointNameLabel: UILabel = {
+        var label = UILabel()
+        label.text = "The Empire State Building"
+        return label
+    }()
 
     lazy var continueButton:UIButton = {
            let button = UIButton(frame: CGRect(x: 0, y: 0, width: 100, height: 100))
@@ -21,7 +33,7 @@ class POIPopUpViewController: UIViewController {
            return button
        }()
        
-       weak var delegate: WaypointConfirmationViewControllerDelegate?
+       //weak var delegate: WaypointConfirmationViewControllerDelegate?
        
        override func viewDidLoad() {
            super.viewDidLoad()
@@ -35,7 +47,7 @@ class POIPopUpViewController: UIViewController {
        }
 
        @objc func continueButtonPressed(_ sender: Any) {
-           delegate?.proceedToNextLegInTour(self)
+           //delegate?.proceedToNextLegInTour(self)
        }
